+++ extensions/prism-launcher/src/index.tsx
@@ -130,6 +130,12 @@ export default function Command() {
                     }}
                   />
                   <Action
+                    title={instance.favorite ? "Remove from Favorites" : "Add to Favorites"}
+                    icon={instance.favorite ? Icon.StarDisabled : Icon.Star}
+                    onAction={() => toggleFavorite(instance.id)}
+                    shortcut={Keyboard.Shortcut.Common.Pin}
+                  />
+                  <Action
                     title="Open Instance Window"
                     icon={"app-window-list-16"}
                     onAction={async () => {
@@ -141,10 +147,21 @@ export default function Command() {
                     }}
                   />
                   <Action
-                    title={instance.favorite ? "Remove from Favorites" : "Add to Favorites"}
-                    icon={instance.favorite ? Icon.StarDisabled : Icon.Star}
-                    onAction={() => toggleFavorite(instance.id)}
-                    shortcut={Keyboard.Shortcut.Common.Pin}
+                    title="Open Minecraft Folder in Finder"
+                    icon={Icon.Finder}
+                    shortcut={{ modifiers: ["cmd"], key: "f" }}
+                    onAction={async () => {
+                      const minecraftPath = path.join(instancesPath, instance.id, "minecraft");
+                      if (await fs.pathExists(minecraftPath)) {
+                        child_process.exec(`open "${minecraftPath}"`);
+                      } else {
+                        child_process.exec(`open "${path.join(instancesPath, instance.id, ".minecraft")}"`);
+                      }
+                      await closeMainWindow({
+                        popToRootType: PopToRootType.Immediate,
+                        clearRootSearch: true,
+                      });
+                    }}
                   />
                 </ActionPanel>
               }